Add try to bluetooth and http calls to keep service up
The pi Zero W sits mirror image to the Wave Plus in a hallway, but it STILL fails to find it more than once a day.  While I was adding that try I went ahead and considered all failure points so the service keeps running.  It does use the Maker API to post an error though.